Fixing Custom endpoint pointing for sub-clients: CallConnection, CallMedia, CallRecording
