+++ sdk/communication/communication-call-automation/src/callAutomationClient.ts
@@ -6,7 +6,6 @@ import { InternalPipelineOptions } from "@azure/core-rest-pipeline";
 import {
   parseClientArguments,
   isKeyCredential,
-  createCommunicationAuthPolicy,
   CommunicationIdentifier,
   CommunicationUserIdentifier,
 } from "@azure/communication-common";
@@ -39,7 +38,7 @@ import {
   PhoneNumberIdentifierModelConverter,
 } from "./utli/converters";
 import { v4 as uuidv4 } from "uuid";
-import { createCallAutomationAuthPolicy } from "./credential/callAutomationAuthPolicy";
+import { createCustomCallAutomationApiClient } from "./credential/callAutomationAuthPolicy";
 import { CallAutomationEventProcessor } from "./eventprocessor/callAutomationEventProcessor";
 import { AnswerCallEventResult, CreateCallEventResult } from "./eventprocessor/eventResponses";
 /**
@@ -69,6 +68,7 @@ export class CallAutomationClient {
   private readonly credential: TokenCredential | KeyCredential;
   private readonly internalPipelineOptions: InternalPipelineOptions;
   private readonly callAutomationEventProcessor: CallAutomationEventProcessor;
+  private readonly endpoint: string;
   /**
    * Initializes a new instance of the CallAutomationClient class.
    * @param connectionString - Connection string to connect to an Azure Communication Service resource.
@@ -112,29 +112,15 @@ export class CallAutomationClient {
     };
 
     const { url, credential } = parseClientArguments(connectionStringOrUrl, credentialOrOptions);
+    this.endpoint = url;
 
     this.credential = credential;
 
     // create event processor
     this.callAutomationEventProcessor = new CallAutomationEventProcessor();
 
-    // read environment variable for callAutomation auth
-    const customEnabled = process.env.COMMUNICATION_CUSTOM_ENDPOINT_ENABLED;
-    const customUrl = process.env.COMMUNICATION_CUSTOM_URL;
-
-    if (customEnabled?.toLowerCase() === "true" && customUrl) {
-      // add custom header for Call Automation auth when flag is true
-      this.callAutomationApiClient = new CallAutomationApiClient(
-        customUrl,
-        this.internalPipelineOptions
-      );
-      const callAutomationAuthPolicy = createCallAutomationAuthPolicy(credential, url);
-      this.callAutomationApiClient.pipeline.addPolicy(callAutomationAuthPolicy);
-    } else {
-      this.callAutomationApiClient = new CallAutomationApiClient(url, this.internalPipelineOptions);
-      const authPolicy = createCommunicationAuthPolicy(credential);
-      this.callAutomationApiClient.pipeline.addPolicy(authPolicy);
-    }
+    // create api client (using custom api endpoint if available)
+    this.callAutomationApiClient = createCustomCallAutomationApiClient(credential, this.internalPipelineOptions, this.endpoint);
 
     this.sourceIdentity = communicationUserIdentifierModelConverter(options.sourceIdentity);
   }
@@ -146,7 +132,7 @@ export class CallAutomationClient {
   public getCallConnection(callConnectionId: string): CallConnection {
     return new CallConnection(
       callConnectionId,
-      this.callAutomationApiClient.endpoint,
+      this.endpoint,
       this.credential,
       this.callAutomationEventProcessor,
       this.internalPipelineOptions
@@ -158,7 +144,7 @@ export class CallAutomationClient {
    */
   public getCallRecording(): CallRecording {
     return new CallRecording(
-      this.callAutomationApiClient.endpoint,
+      this.endpoint,
       this.credential,
       this.internalPipelineOptions
     );
@@ -205,7 +191,7 @@ export class CallAutomationClient {
       };
       const callConnection = new CallConnection(
         callConnectionId,
-        this.callAutomationApiClient.endpoint,
+        this.endpoint,
         this.credential,
         this.callAutomationEventProcessor,
         this.internalPipelineOptions
@@ -339,7 +325,7 @@ export class CallAutomationClient {
       };
       const callConnection = new CallConnection(
         callConnectionId,
-        this.callAutomationApiClient.endpoint,
+        this.endpoint,
         this.credential,
         this.callAutomationEventProcessor,
         this.internalPipelineOptions
