+++ sdk/communication/communication-call-automation/src/callConnection.ts
@@ -3,7 +3,6 @@
 
 import {
   CommunicationIdentifier,
-  createCommunicationAuthPolicy,
 } from "@azure/communication-common";
 import { CallMedia } from "./callMedia";
 import {
@@ -57,6 +56,7 @@ import {
   RemoveParticipantEventResult,
   TransferCallToParticipantEventResult,
 } from "./eventprocessor/eventResponses";
+import { createCustomCallAutomationApiClient } from "./credential/callAutomationAuthPolicy";
 
 /**
  * CallConnection class represents call connection based APIs.
@@ -76,9 +76,7 @@ export class CallConnection {
     eventProcessor: CallAutomationEventProcessor,
     options?: CallAutomationApiClientOptionalParams
   ) {
-    this.callAutomationApiClient = new CallAutomationApiClient(endpoint, options);
-    const authPolicy = createCommunicationAuthPolicy(credential);
-    this.callAutomationApiClient.pipeline.addPolicy(authPolicy);
+    this.callAutomationApiClient = createCustomCallAutomationApiClient(credential, options, endpoint);
     this.callConnectionId = callConnectionId;
     this.callConnection = new CallConnectionImpl(this.callAutomationApiClient);
     this.endpoint = endpoint;
