+++ sdk/communication/communication-call-automation/src/callMedia.ts
@@ -26,7 +26,6 @@ import { CallMediaImpl } from "./generated/src/operations";
 
 import {
   CommunicationIdentifier,
-  createCommunicationAuthPolicy,
   serializeCommunicationIdentifier,
 } from "@azure/communication-common";
 
@@ -55,6 +54,7 @@ import {
 } from "./eventprocessor/eventResponses";
 import { CallAutomationEventProcessor } from "./eventprocessor/callAutomationEventProcessor";
 import { v4 as uuidv4 } from "uuid";
+import { createCustomCallAutomationApiClient } from "./credential/callAutomationAuthPolicy";
 
 /**
  * CallMedia class represents call media related APIs.
@@ -71,9 +71,7 @@ export class CallMedia {
     eventProcessor: CallAutomationEventProcessor,
     options?: CallAutomationApiClientOptionalParams
   ) {
-    this.callAutomationApiClient = new CallAutomationApiClient(endpoint, options);
-    const authPolicy = createCommunicationAuthPolicy(credential);
-    this.callAutomationApiClient.pipeline.addPolicy(authPolicy);
+    this.callAutomationApiClient = createCustomCallAutomationApiClient(credential, options, endpoint);
     this.callConnectionId = callConnectionId;
     this.callAutomationEventProcessor = eventProcessor;
     this.callMedia = new CallMediaImpl(this.callAutomationApiClient);
