+++ sdk/communication/communication-call-automation/src/credential/callAutomationAuthPolicy.ts
@@ -8,6 +8,8 @@ import {
 } from "@azure/core-rest-pipeline";
 import { KeyCredential, TokenCredential, isTokenCredential } from "@azure/core-auth";
 import { createCallAutomationAccessKeyCredentialPolicy } from "./callAutomationAccessKeyCredentialPolicy";
+import { CallAutomationApiClient, CallAutomationApiClientOptionalParams } from "./../generated/src";
+import { createCommunicationAuthPolicy } from "@azure/communication-common";
 /**
  * Creates a pipeline policy to authenticate request based
  * on the credential passed in.
@@ -28,4 +30,35 @@ export function createCallAutomationAuthPolicy(
   } else {
     return createCallAutomationAccessKeyCredentialPolicy(credential, acsUrl);
   }
+}
+
+/**
+ * Creates CallAutomationApiClient for custom endpoint
+ * @hidden
+ *
+ * @param credential - The KeyCredential or TokenCredential.
+ * @param internalPipelineOptions - CallAutomationApiClientOptionalParams if provided.
+ * @param url - ACS url.
+ */
+export function createCustomCallAutomationApiClient(
+  credential: KeyCredential | TokenCredential,
+  internalPipelineOptions: CallAutomationApiClientOptionalParams | undefined,
+  url: string
+): CallAutomationApiClient {
+    // read environment variable for callAutomation auth
+    const customEnabled = process.env.COMMUNICATION_CUSTOM_ENDPOINT_ENABLED;
+    const customUrl = process.env.COMMUNICATION_CUSTOM_URL;
+    let callAutomationApiClient: CallAutomationApiClient;
+
+    if (customEnabled?.toLowerCase() === "true" && customUrl) {
+      // add custom header for Call Automation auth when flag is true
+      callAutomationApiClient = new CallAutomationApiClient(customUrl, internalPipelineOptions);
+      const callAutomationAuthPolicy = createCallAutomationAuthPolicy(credential, url);
+      callAutomationApiClient.pipeline.addPolicy(callAutomationAuthPolicy);
+    } else {
+      callAutomationApiClient = new CallAutomationApiClient(url, internalPipelineOptions);
+      const authPolicy = createCommunicationAuthPolicy(credential);
+      callAutomationApiClient.pipeline.addPolicy(authPolicy);
+    }
+    return callAutomationApiClient;
 }
